requirements.txt
Django==2.2.5
psycopg2-binary==2.8.6
mpld3==0.5.2
biopython==1.78
panoptes-ui==0.2.0
wget==3.2
celery==4.4.7
django-celery-results==2.0.1
flower==0.9.7
celery-progress==0.1.0
altair==4.1.0
pandas==1.2.4
turtle==0.0.1
seaborn==0.12.0
bokeh==2.4.3

